requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
requests==2.31.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
google-search-results==2.4.2
python-multipart==0.0.6
aiohttp==3.9.1
gunicorn==21.2.0